+++ src/egon/data/datasets/power_plants/__init__.py
@@ -16,10 +16,16 @@ from sqlalchemy.orm import sessionmaker
 import geopandas as gpd
 import numpy as np
 import pandas as pd
+
 from egon.data import db
 from egon.data.datasets import Dataset
 from egon.data.datasets.power_plants.pv_rooftop import pv_rooftop_per_mv_grid
 import egon.data.config
+from egon.data.datasets.power_plants.conventional import (
+    select_nep_power_plants,
+    select_no_chp_combustion_mastr,
+    match_nep_no_chp,
+)
 import egon.data.datasets.power_plants.wind_farms as wind_onshore
 import egon.data.datasets.power_plants.wind_offshore as wind_offshore
 import egon.data.datasets.power_plants.pv_ground_mounted as pv_ground_mounted
@@ -47,26 +53,21 @@ class EgonPowerPlants(Base):
 
 
 class PowerPlants(Dataset):
+
     def __init__(self, dependencies):
         super().__init__(
             name="PowerPlants",
-            version="0.0.2",
+            version="0.0.3",
             dependencies=dependencies,
             tasks=(
                 create_tables,
                 insert_hydro_biomass,
-<<<<<<< HEAD
-                 wind_onshore.insert,
-                 pv_ground_mounted.insert,
-                 pv_rooftop_per_mv_grid,
-                 wind_offshore.insert               
-                 assign_weather_data.weather_id,
-=======
+                allocate_conventional_non_chp_power_plants,
                 wind_onshore.insert,
                 pv_ground_mounted.insert,
                 pv_rooftop_per_mv_grid,
+                wind_offshore.insert               
                 assign_weather_data.weather_id,
->>>>>>> features/#330-assign-weather-cell-id-to-power-plants
             ),
         )
 
@@ -520,7 +521,7 @@ def assign_bus_id(power_plants, cfg):
         power_plants.loc[power_plants_hv, "bus_id"] = gpd.sjoin(
             power_plants[power_plants.index.isin(power_plants_hv)],
             mv_grid_districts,
-        ).bus_id
+        ).subst_id
 
     # Assign power plants in ehv to ehv bus
     power_plants_ehv = power_plants[power_plants.voltage_level < 3].index
@@ -589,3 +590,150 @@ def insert_hydro_biomass():
     for scenario in ["eGon2035"]:
         insert_biomass_plants(scenario)
         insert_hydro_plants(scenario)
+
+
+def allocate_conventional_non_chp_power_plants():
+
+    carrier = ["oil", "gas", "other_non_renewable"]
+
+    cfg = egon.data.config.datasets()["power_plants"]
+
+    for carrier in carrier:
+
+        nep = select_nep_power_plants(carrier)
+        mastr = select_no_chp_combustion_mastr(carrier)
+
+        # Assign voltage level to MaStR
+        mastr["voltage_level"] = assign_voltage_level(
+            mastr.rename({"el_capacity": "Nettonennleistung"}, axis=1), cfg
+        )
+
+        # Initalize DataFrame for matching power plants
+        matched = gpd.GeoDataFrame(
+            columns=[
+                "carrier",
+                "el_capacity",
+                "scenario",
+                "geometry",
+                "MaStRNummer",
+                "source",
+                "voltage_level",
+            ]
+        )
+
+        # Match combustion plants of a certain carrier from NEP list
+        # using PLZ and capacity
+        matched, mastr, nep = match_nep_no_chp(
+            nep, mastr, matched, buffer_capacity=0.1, consider_carrier=False
+        )
+
+        # Match plants from NEP list using city and capacity
+        matched, mastr, nep = match_nep_no_chp(
+            nep,
+            mastr,
+            matched,
+            buffer_capacity=0.1,
+            consider_carrier=False,
+            consider_location="city",
+        )
+
+        # Match plants from NEP list using plz,
+        # neglecting the capacity
+        matched, mastr, nep = match_nep_no_chp(
+            nep,
+            mastr,
+            matched,
+            consider_location="plz",
+            consider_carrier=False,
+            consider_capacity=False,
+        )
+
+        # Match plants from NEP list using city,
+        # neglecting the capacity
+        matched, mastr, nep = match_nep_no_chp(
+            nep,
+            mastr,
+            matched,
+            consider_location="city",
+            consider_carrier=False,
+            consider_capacity=False,
+        )
+
+        # Match remaining plants from NEP using the federal state
+        matched, mastr, nep = match_nep_no_chp(
+            nep,
+            mastr,
+            matched,
+            buffer_capacity=0.1,
+            consider_location="federal_state",
+            consider_carrier=False,
+        )
+
+        # Match remaining plants from NEP using the federal state
+        matched, mastr, nep = match_nep_no_chp(
+            nep,
+            mastr,
+            matched,
+            buffer_capacity=0.7,
+            consider_location="federal_state",
+            consider_carrier=False,
+        )
+
+        print(f"{matched.el_capacity.sum()} MW of {carrier} matched")
+        print(f"{nep.c2035_capacity.sum()} MW of {carrier} not matched")
+
+        matched.crs = "EPSG:4326"
+
+        # Assign bus_id
+        # Load grid district polygons
+        mv_grid_districts = db.select_geodataframe(
+            f"""
+        SELECT * FROM {cfg['sources']['egon_mv_grid_district']}
+        """,
+            epsg=4326,
+        )
+
+        ehv_grid_districts = db.select_geodataframe(
+            f"""
+        SELECT * FROM {cfg['sources']['ehv_voronoi']}
+        """,
+            epsg=4326,
+        )
+
+        # Perform spatial joins for plants in ehv and hv level seperately
+        power_plants_hv = gpd.sjoin(
+            matched[matched.voltage_level >= 3],
+            mv_grid_districts[["bus_id", "geom"]],
+            how="left",
+        ).drop(columns=["index_right"])
+        power_plants_ehv = gpd.sjoin(
+            matched[matched.voltage_level < 3],
+            ehv_grid_districts[["bus_id", "geom"]],
+            how="left",
+        ).drop(columns=["index_right"])
+
+        # Combine both dataframes
+        power_plants = pd.concat([power_plants_hv, power_plants_ehv])
+
+        # Delete existing CHP in the target table
+        db.execute_sql(
+            f""" DELETE FROM {cfg ['target']['schema']}.{cfg ['target']['table']}
+            WHERE carrier IN ('gas', 'other_non_renewable', 'oil')
+            AND scenario='eGon2035';"""
+        )
+
+        # Insert into target table
+        session = sessionmaker(bind=db.engine())()
+        for i, row in power_plants.iterrows():
+            entry = EgonPowerPlants(
+                sources={"el_capacity": row.source},
+                source_id={"MastrNummer": row.MaStRNummer},
+                carrier=row.carrier,
+                el_capacity=row.el_capacity,
+                voltage_level=row.voltage_level,
+                bus_id=row.bus_id,
+                scenario=row.scenario,
+                geom=f"SRID=4326;POINT({row.geometry.x} {row.geometry.y})",
+            )
+            session.add(entry)
+        session.commit()